requirements.txt
flask
docker
rich
pydantic
boto3
psycopg2-binary
python-dotenv
requests